Initial commit, beebmod, a JS player for MOD on the BBC Micro.
As of currently it will load a MOD file and start to play basic
parts of it with Amiga-like quality.

As of later, it will play the MOD files on a simulated SN76489 as
driven by the available bandwith on the BBC Micro (aka. beeb).
This will enable fast iteration to try and develop a MOD player for
the beeb that sounds reasonable within the constraints. There are
a lot of different parameters to fiddle with and schemes to try.