Add volume support.
Currently 8 levels of volume as it's not clear how many volume
tables we'd have memory for in a beeb player.

This adds as many problems as it solves.
The challenge with the beeb players is vertical sample resolution,
both in the sample values and in the mapping of sample values to
SN volume outputs. The vertical sample resolution problem is a
headache at the best of times, so trying to play quieter versions
of a given sample only makes it worse.
That said, some samples do respond well to being played quietly,
and the "offset" trick helps.