Generate compressed advance tables.
Compression is a simple bit packing, since the largest byte value is
2 or 3.
We decompress the tables at runtime in 6502.

Doing this dance enables us to use more memory. When loading files,
we need to leave some of the DFS workspace along. Once all files are
loaded, we can decompress into the DFS workspace.