Make the end sample pad length dynamic.
The pad at the end of the sample is to enable sample wrap (or
transition to silence) in a high performance manner. A slight
sample overrun is tolerated for a short amount of time.

Make the pad length configurable as a command line parameter to
sample_adjust.

Change the current pad length to 24 because the highest pitch note
is believed to lead to an overrun of up to 18.