Handle sample page rollover differently.
Currently applies to 15k player.

Gains 5 cycles in every handler block.
The main cost is that handler blocks will run every so slightly
less often, due to the need for a "carry fixup" block to run from
time to time.
This opens the door for a simple effect on one or two of the channels
by inserting a lookup table.